### 2017-07-10 #133 @rflprr
data.world
This package allows users to consume data hosted on data.world. By way of its sibling package (dwapi) it also allows users to invoke data.world API endpoints.

```
Package: data.world
Title: Main Package for Working with 'data.world' Data Sets
Version: 1.1.1
    person("Rafael", "Pereira", email = "rafael.pereira@data.world", role = c("aut", "cre")),
    person("Triet", "Le", email = "triet.le@data.world", role = c("aut")),
    person("Bryon", "Jacob", email = "bryon.jacob@data.world", role = c("aut")))
Description: High-level tools for working with data.world data sets. data.world is a community
    where you can find interesting data, store and showcase your own data and data projects,
    and find and collaborate with other members. In addition to exploring, querying and
    charting data on the data.world site, you can access data via 'API' endpoints and
    integrations. Use this package to access, query and explore data sets, and to
    integrate data into R projects. Visit <https://data.world>, for additional information.
License: Apache License 2.0
    dwapi,
    httr,
    ini
    readr
License: Apache License 2.0
```

https://github.com/datadotworld/data.world-r

[e.g., "data extraction, because the package parses a scientific data file format"]

This package fits under Data Retrieval (and/or, possibly, Data Publication), because it allows access to data on data.world and, also, creation of datasets via API wrappers in the dwapi package.

Data scientists, data analysts, researchers, etc. — any professional or enthusiast working on data projects.

I am not aware of any other R packages designed to work with data.world.
