### 2017-07-11 #133 @rflprr
Hi, @maelle. data.world is an open and community-driven platform. We have built a general purpose tool that serves researchers increasingly well (we are still early stage, but quickly adding functionality to aid reproducibility, for example). Searching the platform you can find several users in academia. Not all their datasets are public, so it's hard to give you an idea for the breath of scientific example. Here is one example of a user who is a PhD student and who have created a few public datasets: https://data.world/jordanageorge?relationship=owner. I hope this helps.
