### 2017-07-18 #133 @sckott
Doing my editor checks now, sorry for the delay ...
