### 2017-07-18 #133 @sckott
HI @rflprr

What was the reasoning for splitting out `dwapi-r` ? I'm not saying I disagree with the split, just asking why.  I think I'm leaning towards asking reviewers to review the pkg submitted, but also to at least have a look at `dwapi-r` since it does some heavy lifting esp. wrt HTTP requests.
