### 2017-07-19 #133 @rflprr
Hi, @sckott. Our plan is to provide wrappers for all our API endpoints. That's an ever growing list of low level operations that would pollute the namespace of our main library, which is intended to always expose a very limited set of high-level and easy to use operations. In short, we separated so there is more clear separation of concerns and more cohesive namespaces to better serve what we believe are two distinct groups of users 1) those who need access to low-level APIs and value flexibility, and 2) the vast majority who only care about a few high-level operations and value ease of use.
