### 2017-07-20 #133 @sckott
@rflprr  Okay. Seems the same goal could be done while having the lower level methods in this pkg, but just not exported to the user. - But if there's use cases for a certain kind of user (power user) to use just the low level methods in the other pkg that makes sense.  I'd like to have reviewers at least take a brief  look at `dwapi-r` in addition to this pkg, okay?
