### 2017-07-20 #133 @rflprr
@sckott, reviewing `dwapi-r` makes a lot of sense. Please go ahead and let me know if you find any issues.
