### 2017-07-25 #133 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @rflprr  !

Currently seeking reviewers. It's a good fit and not overlapping.

* `goodpractice::gp()` output is below for reviewers to consider - feel free to run it yourself as well

```r
It is good practice to

    R/data.world-defunct.R:65:NA
    R/data.world-defunct.R:67:NA
    R/data.world-defunct.R:68:NA
    R/data.world-defunct.R:70:NA
    R/data.world-defunct.R:71:NA
    ... and 55 more lines
```

For ``

```r
It is good practice to

    R/add_file.R:65:NA
    R/add_file.R:95:NA
    R/add_file.R:97:NA
    R/add_file.R:98:NA
    R/add_file.R:99:NA
    ... and 162 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R/dataset_summary_response.R:41:15
    R/table_schema_response.R:28:15

  ✖ fix this R CMD check NOTE: Namespace in Imports field not
    imported from: ‘xml2’ All declared Imports should be used.
  ✖ fix this R CMD check NOTE: Found the following hidden
    files and directories: .lintr These were most likely included in
    error. See section ‘Package structure’ in the ‘Writing R
    Extensions’ manual.
```

---

Reviewers:
Due date:
