### 2017-08-04 #133 @MarkEdmondson1234
## Package Review - data.world-r (dwapi-r)

### Summmary

Overall the package is a nice addition, allow access to a crowd-sourced variety of data that would be useful in many projects.  The richness of the datasets will depend largely on how popular the website `data.world` becomes, but already there are some good examples for teaching and in various sectors.  This package's success is tied to the success of that endeavour, to which I encourage and wish well.

The line between `data.world` and `dwapi` is a little blurry, and I do question the need to have seperate packages given as in the short time I got to use `data.world` I was already needing to dip into the `dwapi` package to get what I needed, with the added confusion of where documentation should sit etc.  I would suggest either to export/provide wrappers for all the functions that are useful in `data.world`, or to merge it all into `dwapi`, particularly as the documentation of `data.world` is a little sparse.  Myself I am inclined to use `dwapi` directly as perhaps other R users will too, but if the documentation was all in `data.world` with the most common useful functions (such as metadata requests) that could change.  If worried about complication of functions available, more use of `@noRd` looks to solve that issue.

### Review template

I could install with no problems following the README instructions.

The "Hello World" example worked quickly and well.

The vignettes were the `quickstart` and `query`.  The Quickstart worked getting youto your first query

### quickstart

I followed the first vignette as suggested via `vignette("quickstart", package = "data.world")`

The installation instructions offer three options (runtime, environment variables, runtime), but only include an example for one (via config file).

The help files in `?save_config` and `?set_config` could detail more about what is happening as well.

The start guide states to not keep your API token in code, but then only gives examples where it is in code.  I manually setup a environment argument in my `.Renviron` for use, but that assumes users know about how to do that (as say explained in [httr's docs](https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html)) but other R users may not.

I did it like this:

```r
saved_cfg <- data.world::save_config(Sys.getenv("DATADOTWORLD"))

data.world::set_config(saved_cfg)
```

...but its not clear if this is a good way to handle it or how `save_config` or `set_config` works as I saw no files in my directory.

Now I inspect the code I see it expects `DW_AUTH_TOKEN` as the environment file - but this needs to be documented somewhere, and the examples updated to show its use.

The API token is also very long, is this necessary?  Compare the GitHub personal access token and this packages:

```r
nchar(Sys.getenv("DATADOTWORLD"))
# [1] 340
nchar(Sys.getenv("GITHUB_PAT"))
# [1] 40
```

### query

This vignette went into a lot more detail on how to query data, with some good links to further documentation.

All the SQL examples worked, but the `SPARQL` SELECT example returned no results:

```r
assists_vs_height <- data.world::qry_sparql(paste(
  "BASE <http://data.world/jonloyens/an-intro-to-dataworld-dataset/> ",
  "PREFIX t: <DataDotWorldBBallTeam.csv/DataDotWorldBBallTeam#> ",
  "PREFIX s: <DataDotWorldBBallStats.csv/DataDotWorldBBallStats#> ",
  "SELECT ?name ?height ?assists WHERE { ",
  "  ?pt t:Name ?name . ",
  "  ?ps s:Name ?name . ", # the join column
  "  ?pt t:Height ?height . ",
  "  ?ps s:AssistsPerGame ?assists . ",
  "} ",
  "ORDER BY DESC (?assists)", sep = "\n"
))

data.world::query(assists_vs_height, intro_ds)
# A tibble: 0 x 3
# ... with 3 variables: name <chr>, height <chr>, assists <chr>
```

And the parameterized example errored:

```r
assists_greater_than <- data.world::qry_sparql(paste(
  "BASE <http://data.world/jonloyens/an-intro-to-dataworld-dataset/> ",
  "PREFIX t: <DataDotWorldBBallTeam.csv/DataDotWorldBBallTeam#> ",
  "PREFIX s: <DataDotWorldBBallStats.csv/DataDotWorldBBallStats#> ",
  "SELECT ?name ?height ?assists WHERE { ",
  "  ?pt t:Name ?name . ",
  "  ?ps s:Name ?name . ", # the join column
  "  ?pt t:Height ?height . ",
  "  ?ps s:AssistsPerGame ?assists . ",
  "  FILTER(?assists > $v1) ",
  "} ",
  "ORDER BY DESC (?assists)", sep = "\n"
))
assists_greater_than$params <- c("$v1" = 10)
data.world::query(assists_greater_than, intro_ds)
# No encoding supplied: defaulting to UTF-8.
# Error in dwapi::sparql(dataset = dataset, query = qry$query_string, query_params = qry$params) :
#   Error: Unbound variable used in filter: $v1   (line 9, column 21)
```

There was a helpful list of defunct functions, I should do that for my packages too :)

There was also a helpful intro help file under `?data.world`

The help files for `cfg`, `cfg_env`, and `cfg_saved` looked to be internal functions - could they be removed from the function list via `#' @noRd`?  Similarly the methods for `query` and `set_config` since they all point at the same documentation.

The documentation for the main useful functions such as `qry_sparql` and `qry_sql` is very sparse, some more detail and at least a link to the documentation that is linked in the vignettes would be useful.

The most important functions in the package (as judged by the vignette metions) do not have examples in their help files yet - namely `qry_sparql` and `qry_sql` - copying over the (working) examples from the vignettes in `\dontrun{}` would be helpful for day to day use.

There are no contributing guidelines yet.

The DESCRIPTION has a URL and BugReports to GitHub, as well as the authors/maintainers via `Authors@R`

The Description looks like it needs to be line-breaked at 80 characters more, as it did not display correctly compared with other packages in the Help pane.

The tests covered a good range and used mock testing to quickly pass on the local machine.

* The functions under rOpenSci guidelines would have some kind of `object_verb` structure, such as `dw_qry_sql`, `dw_set_config`, assuming `dw_` was your chosen object.
* In the `README` the Installation and Configuration section should be higher in the page before the demo.
* The rOpenSci review badge is not yet present
* A bit more description of what `Data World` does as an API service overall, as described in the rOpenSci submission.
* A Code of Conduct could be added
* The `NEWS.md` could have more details such as package name, version and date.  It should also be in descending date order.
* The package versioning usually also follows what is released on CRAN, not revisions before that (for development versions off CRAN, the suggestion is to use `1.1.1.9000` as a way to distinguish them)  See [here](http://r-pkgs.had.co.nz/release.html) for more details.
* In `dwapi`, `rjson` is used instead of the guidelines `jsonlite` - example code: [`dwapi::add_files_by_source`](https://github.com/datadotworld/dwapi-r/blob/master/R/add_files_by_source.R#L71)

### Code review - data.world

`data.world` moves a lot of the work to the supporting `dwapi` package, so only minor suggestions here

#### structure()

The below code could use the existing `structure()` function, and there is no need for the `return()`:

e.g.

```r
## present function
qry_sql <- function(query_string, params = NULL) {
  me <- list(query_string = query_string, params = params)
  class(me) <- "qry_sql"
  return(me)
}

## suggested
qry_sql <- function(query_string, params = NULL) {
  structure(
    list(query_string = query_string, params = params),
    class = "qry_sql"
  )
}
```

Repeat for `qry_sparql`, `cfg`, `cfg_env`, `cfg_saved` etc.

#### config ini files

Only after examination of the code did I find the package is writing `.ini` files to my system without my knowledge. e.g. `getOption("dw.config_path") = "/Users/mark/.dw/config"` - I would suggest at the very least to document this somewhere and/or a message stating a file is being written to my system.  The config file is also written outside of my working directory, which is unexpected.  I would prefer it writes to the working directory in a visible file, and I can choose the option myself if it is to be written elsewhere on my system, especially if this is a file that is created/saved on loading the package via `.onAttach`.

#### discovery queries

Its a little confusing what should `data.world` be used for compared with `dwapi` - for example I was looking for the functionality of `dwapi::list_tables()` in the `data.world` package, which I think should have at least a wrapper for or re-exported or something.

### Code review - dwapi

#### linter issue?

The formatting for function arguments looked like incorrect indentation to me, and RStudio agreed when I did `CMD+I` on the lines:

```r
# before
file_create_request <- function(file_name,
  url,
  description = NULL,
  labels = NULL) {
  if (file_name == "" | url == "") {
    stop("name and url required")
  }

  me <-
    list(name = file_name, source = file_source_create_request(url))
  if (!is.null(description)) {
    me$description <- description
  }
  if (!is.null(labels)) {
    me$labels <- labels
  }

  class(me) <- "file_create_request"
  return(me)
}

# after CMD+I
file_create_request <- function(file_name,
                                url,
                                description = NULL,
                                labels = NULL) {
  if (file_name == "" | url == "") {
    stop("name and url required")
  }

  me <-
    list(name = file_name, source = file_source_create_request(url))
  if (!is.null(description)) {
    me$description <- description
  }
  if (!is.null(labels)) {
    me$labels <- labels
  }

  class(me) <- "file_create_request"
  return(me)
}
```

#### using structure()

Again `structure()` could be used to make the classes, and a helper class to get rid of `NULL`s from a list will let you make neater objects:

```r
# before
file_create_or_update_request <- function(file_name,
                                          url = NULL,
                                          description = NULL,
                                          labels = NULL) {
  if (file_name == "") {
    stop("name is required")
  }

  me <- list(name = file_name)
  if (!is.null(url)) {
    me$source <- file_source_create_or_update_request(url)
  }
  if (!is.null(description)) {
    me$description <- description
  }
  if (!is.null(labels)) {
    me$labels <- labels
  }

  class(me) <- "file_create_or_update_request"
  return(me)
}

# after

# function to remove NULLs from lists.
compact <- function(x) {
  null <- vapply(x, is.null, logical(1))
  x[!null]
}

file_create_or_update_request <- function(file_name,
                                          url = NULL,
                                          description = NULL,
                                          labels = NULL) {
  if (file_name == "") {
    stop("name is required")
  }

  me <- structure(
    list(
      name = file_name,
      # may need to alter to return NULL if passed NULL
      source = file_source_create_or_update_request(url),
      description = description,
      labels = labels
    ),
    class = "file_create_or_update_request"
  )

  compact(me)

}
```

#### growing vectors

The method of adding to requests doesn't look like R standards since its using `your_vector[[length(your_vector) + 1]] <- my_addition` rather than `c(your_vector, my_addition)`

Also why is `dwapi::` used here within the `dwapi` package?  I guess it got migrated and forgot its internal now.

```r
# before
add_file.dataset_create_request <-
  function(request,
           name,
           url,
           description = NULL,
           labels = NULL) {
    existing_files <- request$files
    # O(N) ?
    existing_files[[length(existing_files) + 1]] <-
      dwapi::file_create_request(
        name, url, description = description, labels = labels)
    request$files <- existing_files
    request
  }

# after
add_file.dataset_create_request <-
  function(request,
           name,
           url,
           description = NULL,
           labels = NULL) {

    request$files <- c(request$files,
                       file_create_request(name,
                                           url,
                                           description = description,
                                           labels = labels))
    request
  }
```

#### Using httr::RETRY

You may want to consider using `httr::RETRY` instead of `httr::POST` for API requests, to mitigate against network errors. [Example it could be here](https://github.com/datadotworld/dwapi-r/blob/master/R/add_files_by_source.R#L69)

#### Creating a generic API fetch function

A lot of the API requests are constructing their own `httr::POST`etc. requests with repeated information such as auth, user agent etc.  This results in lots of repeated code, so I suggest a generic "fetch the API with these arguments" function be created, and then pass arguments to that.

e.g something like:

```
dw_do_api <- function(verb, path_url, query_param = NULL, body = NULL){

  base_url <- getOption("dwapi.api_url")
  api_url <- paste0(base_url, path_url)

  if(!is.null(query_param)){
    api_url <- paste0("?", paste0(query_param, collapse="&"))
  }

  response <-
        httr::RETRY(verb,
                    url = api_url,
                    httr::add_headers(Authorization = sprintf("Bearer %s", auth_token())),
                    httr::user_agent(user_agent()),
                    body = body
                    )

  parse_success_or_error(response)
}
```

Then API functions can use it like:

```r
# before
delete_files <-
  function(dataset, file_names) {
    query_param <-
      lapply(file_names, function(name)
        sprintf("name=%s", name))
    if (length(query_param) == 0) {
      print("empty names input = no-op")
      api_url <-
        paste(
          sprintf(
            "%s/datasets/%s/files?", getOption("dwapi.api_url"),
            extract_dataset_key(dataset)),
          paste0(query_param, collapse = "&"),
          sep = ""
        )
      auth <- sprintf("Bearer %s", auth_token())
      response <-
        httr::DELETE(
          api_url,
          httr::add_headers(Authorization = auth),
          httr::user_agent(user_agent())
        )
      ret <- parse_success_or_error(response)
      ret
    }
  }

# after
delete_files <-
  function(dataset, file_names) {
    query_param <-
      lapply(file_names, function(name)
        sprintf("name=%s", name))
    if (length(query_param) == 0) {
      print("empty names input = no-op")
      dw_do_api("DELETE",
                path_url = sprintf("/datasets/%s/files", extract_dataset_key(dataset)),
                query_param = query_param)
    }
  }
```

Repeat as necessary for all the other API requests.  You could also add the response status code test for `200` and parsing out the error message if incorrect to the `dw_do_api` function which should mean a lot less code needed overall.  For instance, swapping from `rjson` to `jsonlite` at the moment will mean a lot of refactoring, which could be avoid by making a centralised fectch/parse function.  You can also then do things like turn on `verbose()` for checking HTTP requests more easily.
