### 2017-08-09 #133 @rflprr
@sckott and @MarkEdmondson1234, first, thank you. Second, I'm unsure about the process at this point. Lot's of good feedback here, but no sense of severity. With limited resources, I'd like to limit any changes to a minimum and continue to implement improvements over time. What is your expectation at this point?

The main point I would like to address is why data.world and dwapi. We are serving two audiences with those packages, those looking for high-level and low entry barrier features and those looking for low-level and maximum flexibility, respectively. I want to make them as easy to learn and their features easy to discover for the right users and that's why I opted for two packages (two namespaces) and to export all relevant functions from each. data.world is sparse right now because it's still in its infancy we want to be judicious about what features we add to it. Our intent is for it to contain functionality that is most useful in the context of interactive programming (RStudio, R Notebooks, etc). `dwapi`, on the other hand, is more full-featured because less thought goes into its design. It will simply evolve along-side our REST API and will be better suited for scripts, headless apps, and web apps (e.g. Shiny).

If you object to this approach or see any severe implementation/conformance flaws, let me know. Otherwise, I'll create issues on our repo for everything you reported and would like to address them over time.

Thanks again!
