### 2017-08-09 #133 @sckott
@rflprr at this point I think it's best to wait for the 2nd review to come in to have all the reviewer comments. Once the 2nd review is in, I can help think about which changes may be most important.
