### 2017-08-23 #133 @nistara
The `data.world` package is mainly for retrieving data from the data.wprd website, whereas the `dwapi` package provides greater functionality like updating and uploading data, editing metadata, etc. It might be helpful to combine the two together.

Overall I think these are very useful packages which will encourage the use and analyses of data from [data.world](https://data.world)!

I suggest giving a little more information about the `data.world` website in the README as well; this would help orient new users and increase interest in the package as well as `data.world`. Perhaps also shift the installation and configuration above the getting started example. There's a slight redundancy in suggesting the checking out of the "quickstart" vignette.

I could install the development version of the package without any problem.

The vignettes were helpful in understanding the core functionalities of the package.

**quickstart**

Does not mention how configuration can be provided via environment variables and at runtime. Tangentially related, when I signed up with [data.world](https://data.world/settings/advanced), there was a typo for "collaborators" in the getting started screen.

**query**

The `query` vignette was useful and it's nice that the examples are based on the same dataset and links to more information on SQL and SPARQL are provided. The SPARQL examples don't work as expected however:

**SELECT queries**
No data is returned with this example:

```

assists_vs_height <- data.world::qry_sparql(paste(
    "BASE <http://data.world/jonloyens/an-intro-to-dataworld-dataset/> ",
    "PREFIX t: <DataDotWorldBBallTeam.csv/DataDotWorldBBallTeam#> ",
    "PREFIX s: <DataDotWorldBBallStats.csv/DataDotWorldBBallStats#> ",
    "SELECT ?name ?height ?assists WHERE { ",
    "  ?pt t:Name ?name . ",
    "  ?ps s:Name ?name . ", # the join column
    "  ?pt t:Height ?height . ",
    "  ?ps s:AssistsPerGame ?assists . ",
    "} ",
    "ORDER BY DESC (?assists)", sep = "\n"
))

data.world::query(assists_vs_height, intro_ds)

# A tibble: 0 x 3
# ... with 3 variables: name <chr>, height <chr>, assists <chr>

```

**Parameterized SPARQL queries**
Returns an error:

```
assists_greater_than <- data.world::qry_sparql(paste(
    "BASE <http://data.world/jonloyens/an-intro-to-dataworld-dataset/> ",
    "PREFIX t: <DataDotWorldBBallTeam.csv/DataDotWorldBBallTeam#> ",
    "PREFIX s: <DataDotWorldBBallStats.csv/DataDotWorldBBallStats#> ",
    "SELECT ?name ?height ?assists WHERE { ",
    "  ?pt t:Name ?name . ",
    "  ?ps s:Name ?name . ", # the join column
    "  ?pt t:Height ?height . ",
    "  ?ps s:AssistsPerGame ?assists . ",
    "  FILTER(?assists > $v1) ",
    "} ",
    "ORDER BY DESC (?assists)", sep = "\n"
))

assists_greater_than$params <- c("$v1" = 10)
data.world::query(assists_greater_than, intro_ds)

# No encoding supplied: defaulting to UTF-8.
# Error in dwapi::sparql(dataset = dataset, query = qry$query_string, query_params = qry$params) :
  # Error: Unbound variable used in filter: $v1   (line 9, column 21)
```

	- some more details in the function descriptions would be great. As of now the descriptions for the functions are the same as their titles.

	- `save_config()` has an argument, `profile`, which is defined as the configuration profile. It doesn't explain what the default profile is, or which other options exist. I think these profiles are the same as the ones listed in `set_config()`, however they should be explained here as well.

	- The README mentions the following, however most of the functions, e.g. `query()`, `qry_sql()` do not provide information on the `dataset` parameter.

	> Notice that dataset is parameter required by most functions and can be provided in two formats:
	>
	>    URL: "https://data.world/jonloyens/an-intro-to-dataworld-dataset"
	>    Path: "jonloyens/an-intro-to-dataworld-dataset"

	- Based on the above points, it might be useful to integrate the information provided in the README, vignettes and the function documentation. Also make the function documentation more detailed than it currently is (including adding more examples).

	- `qry_sql()` and `qry_sparql()` do not have examples in their respective documentation.

	- The example provided for `query()` gives an error, probably because it should be `qry_sql` instead of `sql`:

	```
	sql_stmt <- data.world::sql("SELECT * FROM Tables")
	query_results_df <- data.world::query(
	sql_stmt, "jonloyens/an-intro-to-dataworld-dataset")

	# Error: 'sql' is not an exported object from 'namespace:data.world'

	```

	- There are no contribution guidelines

	- `URL` and `BugReports` links are provided in the DESCRIPTION, and so is the  `Maintainer` (autogenerated via `Authors@R`)

	- `object_verb()` naming scheme could be considered, although function names might get too long when differentiating between sql and sparql.

	- As mentioned before, the README should include citation information, and also consider restructing to first, `Installation instructions`, and then the `Getting started`/`Brief demonstration usage`. More information about data.world itself should be also be provided in the README.

	- The `news.md` file should contain the package name and dates of release. Consider using the section headers `NEW FEATURES`, `MINOR IMPROVEMENTS`, `BUG FIXES`, `DEPRECATED` AND `DEFUNCT` while providing information about each release. See [here](https://github.com/ropensci/onboarding/blob/master/news_template.md).

	- The package passed `R CMD check/devtools::check()`

	- Consider adding a coverage badge to the README

	- `rjson` is used instead of `jsonlite` in the following places:

	  ```
	  ./R/add_files_by_source.R:71:        body = rjson::toJSON(file_batch_update_req),
	  ./R/create_dataset.R:46:      body = rjson::toJSON(create_dataset_req),
	  ./R/create_dataset.R:57:      create_dataset_response(rjson::fromJSON(httr::content(
	  ./R/create_dataset.R:64:      error_message(rjson::fromJSON(httr::content(
	  ./R/get_dataset.R:42:      rjson::fromJSON(httr::content(
	  ./R/get_dataset.R:50:      error_message(rjson::fromJSON(httr::content(
	  ./R/get_table_schema.R:43:      rjson::fromJSON(httr::content(
	  ./R/get_table_schema.R:51:      error_message(rjson::fromJSON(httr::content(
	  ./R/list_tables.R:47:      rjson::fromJSON(httr::content(
	  ./R/list_tables.R:54:      error_message(rjson::fromJSON(httr::content(
	  ./R/replace_dataset.R:39:        body = rjson::toJSON(dataset_replace_req),
	  ./R/update_dataset.R:44:        body = rjson::toJSON(dataset_update_req),
	  ./R/update_table_schema.R:39:    body = rjson::toJSON(table_schema_update_req),
	  ./R/utils.R:57:        success_message(rjson::fromJSON(httr::content(
	  ./R/utils.R:64:        error_message(rjson::fromJSON(httr::content(
	  ```

	 - `goodpractice::gp()` for `dwapi` says:

	 ```
	  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
          1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
          result 1:0 if the expression on the right hand side is zero. Use
          seq_len() or seq_along() instead.
	 ```

    However looking at the respective lines, the authors do check whether length is > 0:

	 ```
	  if (length(structure$files) > 0) {
	    for (i in 1:length(structure$files)) {
		me$files[[i]] <-
           file_summary_response(structure$files[[i]])
	    }
	  }
	 ```

	 - I agree with @MarkEdmondson1234 regarding indentation of code in `dwapi`, and this occurs beyond the example he provided, for e.g. even in `add_file_by_source()`.

	 - Maybe add the `.lintr` file to `.Rbuildignore`?

	 - Spelling errors in `dwapi`:

	 ```
	  devtools::spell_check()
	  WORD           FOUND IN
	  begining       dwapi.Rd:16
	  compability    list_tables.Rd:21, sparql.Rd:21
	  reponse        create_dataset_response.Rd:5,16
	 ```
  - I'm unfamiliar with `structure()` as  @MarkEdmondson1234 suggests (thank you - learnt something new!). I do however see differences in the way objects are returned. Sometimes `return()` is used, and other times it's not. It's probably better practice to be consistent about it (and use `return()` only if needed).

	 ```
	  #' Return the current data.world version.
	  #' @return Current package version.
	  #' @keywords internal
	  sdk_version <- function() {
		is.nothing <- function(x)
		  is.null(x)
		# nolint start
		if (!is.nothing(utils::sessionInfo()$otherPkgs$data.world)) {
		  ret <- utils::sessionInfo()$otherPkgs$dwapi$Version
		  ret <- "X.X.X"
		}
		# nolint end
		ret
	  }

	  #' Return the data.world user-agent.
	  #' @return User-agent string.
	  #' @keywords internal
	  user_agent <- function() {
		ret <- sprintf("data.world-R - %s", sdk_version())
		ret
	  }

	  #' Extract the dataset key from URL or as provided.
	  #' @param tentative_key key or URL.
	  #' @return dataset key extracted form URL or as provided.
	  #' @keywords internal
	  extract_dataset_key <- function(tentative_key) {
		url <- httr::parse_url(tentative_key)
		return(url$path)
	  }
	 ```

	- `data.world` vs `dwapi`
	  It looks like the authors have their reasons for having these two different packages, and while I understand that `data.world` and `dwapi` could be aimed at users with different skill levels, functions that do the same thing across both packages should be called similarly. For e.g. `data.world::query()` basically does the same thing as `dwapi::sql()` or `dwapi::sparql()`, the difference being that queries in `data.world::query()` are passed in with `qry_sql()` or `qry_sparqll()`. This could get confusing for users trying to reconcile the two packages or switch between them.
