### 2017-09-07 #133 @rflprr
@sckott While constructive and certainly helpful, I'm not seeing anything in the feedback that appears to be severe/urgent. If you do, calling those things out would be the most helpful thing at this point.

Thank you very much, @MarkEdmondson1234 and @nistara.

Off-topic: If you know good R engineers who we could hire on a freelance basis to help us improve and continue to develop these libraries, I would appreciate introductions.
