### 2017-09-07 #133 @sckott
@rflprr can you clarify what you mean by "I'm not seeing anything in the feedback that appears to be severe/urgent"?

the way our process works is we want submitters to respond to reviewers, making changes or if they don't think they should change something then give a reason why not.
