### 2017-09-07 #133 @sckott
to clarify:

The purpose of this review is to improve the package along many lines (usability, performance, documentation, maintainability). The first people you need to convince of what is or is not necessary are the reviewers. Here's a good example of responding to reviewers (though it doesn't all have to be in one comment): https://github.com/ropensci/onboarding/issues/93#issuecomment-291005794

If you and a reviewer disagree about the necessity of something, then, an editor can weigh in (e.g., https://github.com/ropensci/onboarding/issues/127#issuecomment-324783498)

While we understand that there is considerable effort in updating a package post-review, we think it is worth it.
