### 2017-10-12 #133 @sckott
@rflprr we do expect responses to reviewers, you don't have to respond to every single item they raised, but no response is not allowed.

Please get your responses to reviewers in within 3 weeks. if not, we'll close the issue, and you're welcome to resubmit at a later point if you like
