### 2017-10-16 #133 @rflprr
@sckott sorry for the communication hiatus (it's been busy period and we are resource constrained).

The reviewers did a very thorough job. Reading though both reviews I see many excellent suggestions, but none that would significantly improve (or de-risk) the user experience in the immediate term (in other words, none that appear to be critical). Most importantly, all suggestions can be implemented in a backwards compatible way over time.

I cannot commit to making these changes at the moment, so we can proceed with the packages in their current stage or pause/abort the onboarding process, if you think that is the right thing to do.

Additionally, we intend to continue to develop these packages if I can find help. I was serious when I asked for references for free-lance developers. If you know good ones and could get me in touch with someone, I would certainly appreciate any intros.
