### 2017-12-27 ca0a74f
Address compatibility issues with testthat 2.0.0; fix minor DESCRIPTION issue w dup keys
