### 2017-12-28 0423734
Initial rough-in of insight addin
