### 2017-12-28 350bfa9
Finish initial version of insight add-in
