### 2017-12-29 471180a
Qualify renderImage() with package
