### 2017-12-29 datadotworld/data.world-r#28 @rflprr
Address feedback from rOpenSci team
Review comments at https://github.com/ropensci/onboarding/issues/133

Comment on all suggestions, indicating:
1. Which ones we believe substantially improve our package and will implement
2. Which ones we believe are not critical and will defer

Create GitHub issues for all in #1.
