### 2017-12-29 eb4f276
Add basic test coverage for insight addin
