### 2017-12-30 d93a1c2
Remove covr exclusion for addin; factor out addin project filtering
