### 2018-01-06 1a9121d
Implement designer feedback
