### 2018-01-06 cc39609
Fix issue running addin without dw packages attached; cleaned up .onLoad/.onAttach functions
