+++ man/set_dw_config.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/zzz.R
+\name{set_dw_config}
+\alias{set_dw_config}
+\title{Apply configuration from file or envvars and load dwapi}
+\usage{
+set_dw_config()
+}
+\description{
+Apply configuration from file or envvars and load dwapi
+}
+\keyword{internal}
