### 2018-01-08 2834b68
Suppress warning message on package load if no config file
