### 2018-01-08 2bd457a
Adjust font and image sizes per designer feedback
