### 2018-01-08 a8ad32d
Further designer tweaks to proportions, centering, etc.
