### 2018-01-09 47b17c5
Tweak text, dependencies and coverage
