### 2018-01-09 aee574f
Address lintr warning
