### 2018-01-09 8236e7a
Clear CMD CHECK issues, improve test coverage
