### 2018-01-12 b39a4c9
Prepare for CRAN release
