+++ NEWS.md
@@ -1,8 +1,14 @@
-# 1.1.0
+# 1.2.0
 
-* Delete all defunct and deprecated functions
-* First attempted CRAN release
+* Introduce "Add Insight" add-in
+* Improve reloading of saved API tokens
+* Address compatibility issues with `testthat` 2.0
 
 # 1.1.1
 
 * Address requests from CRAN reviewers related to 1.1.0
+
+# 1.1.0
+
+* Delete all defunct and deprecated functions
+* First attempted CRAN release
