### 2018-01-12 ec66e74
Initial version of Insight RStudio add-in (#27)
* Address compatibility issues with testthat 2.0.0; fix minor DESCRIPTION issue w dup keys

* Initial rough-in of insight addin

* Added copyright notice

* Added copyright notice

* Finish initial version of insight add-in

* Qualify renderImage() with package

* Apply data.world css

* Add basic test coverage for insight addin

* Remove covr exclusion for addin; factor out addin project filtering

* Implement designer feedback

* Fix issue running addin without dw packages attached; cleaned up .onLoad/.onAttach functions

* Suppress warning mes
[…]
