+++ R/insight-addin.R
@@ -1,0 +1,225 @@
+"data.world-r
+Copyright 2018 data.world, Inc.
+
+Licensed under the Apache License, Version 2.0 (the \"License\");
+you may not use this file except in compliance with the License.
+
+You may obtain a copy of the License at
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an \"AS IS\" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+implied. See the License for the specific language governing
+permissions and limitations under the License.
+
+This product includes software developed at data.world, Inc.
+https://data.world"
+
+#' Driver function for data.world Insight Add-in
+#' @keywords internal
+#' @importFrom dwapi get_projects_user_own get_projects_user_contributing
+#' @importFrom grDevices dev.copy dev.list dev.off png
+#' @importFrom graphics mtext plot.new
+# nocov start
+add_insight_addin <- function() {
+
+  MAX_PROJECT_TITLE_LABEL <- 34
+
+  api_token <- getOption("dwapi.auth_token")
+
+  if (is.null(api_token)) {
+    #nolint start
+    stop(paste0("API authentication must be configured. ",
+                "To configure, use data.world::set_config(save_config(auth_token = \"YOUR API TOKEN\")) or ",
+                "dwapi::configure()"))
+    #nolint end
+  }
+
+  project_list <- insight_project_filter(
+    c(get_projects_user_own()$records,
+      get_projects_user_contributing()$records))
+
+  project_choice_list <- sapply(
+    USE.NAMES = FALSE, project_list, function(project) {
+    paste0(project$owner, "/", project$id)
+  }
+  )
+
+  names(project_choice_list) <- sapply(
+    USE.NAMES = FALSE, project_list, function(project) {
+    ret <- paste0(project$owner, "/", project$title)
+    if (nchar(ret) > MAX_PROJECT_TITLE_LABEL) {
+      ret <- paste0(substr(ret, 1, MAX_PROJECT_TITLE_LABEL), "...",
+                    collapse = "")
+    }
+    ret
+  }
+  )
+
+  ui <- miniUI::miniPage(
+
+    shiny::includeCSS(system.file("dw-bootstrap.css", package = "data.world")),
+
+    shiny::tags$head(
+      shiny::tags$style(
+        type = "text/css",
+        paste(".form-group {font-size: 14px;}",
+              ".form-control {font-size: 14px;}",
+              "#done {line-height: 1rem;}",
+              "#cancel {line-height: 1rem;}",
+              "#title {font-size: 14px;}",
+              "#description {font-size: 14px; min-height: 110px;}",
+              "div.gadget-title {background-color: #fff}",
+              "div.gadget-content {background-color: #fff}",
+              sep = "\n"
+        )
+      ),
+      shiny::tags$link(
+        rel = "stylesheet", type = "text/css",
+        href = "http://fonts.googleapis.com/css?family=Lato:300,300i,400,400i,700,700i") # nolint
+    ),
+
+    miniUI::gadgetTitleBar("Create a new insight",
+                           right = miniUI::miniTitleBarButton("done",
+                                                      "Create insight",
+                                                      primary = TRUE)),
+    miniUI::miniContentPanel(
+      shiny::fillRow(
+      shiny::column(12,
+        shiny::imageOutput("thumb", height = NULL, width = "90%")),
+        shiny::column(12,
+                      shiny::selectInput("project", "Project:",
+                                         choices = c("", project_choice_list)),
+                      shiny::textInput("title", "Title:"),
+                      shiny::textAreaInput("description", "Description:",
[…]
