### 2018-02-23 7e33ed5
Improve CircleCI configuration
