### 2018-02-23 8400304
Fix config.yml syntax
