### 2018-04-03 4287f80
Inline dw_set_config()
