### 2018-04-04 0bceb62
Restore prior .onAttach() behavior
