+++ man/is_attached.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/zzz.R
+\name{is_attached}
+\alias{is_attached}
+\title{Determine if library is already attached}
+\usage{
+is_attached(x)
+}
+\description{
+Determine if library is already attached
+}
+\keyword{internal}
