### 2018-04-04 ebd7ca0
Make dwapi dependency explicit (#33)
* Make dwapi package dependency explicit (Issue #31)

* Restore prior .onAttach() behavior
