### 2018-04-04 b44c881
Ready for CRAN - 1.2.1
