### 2018-04-04 2f0bea2
Fix insight addin bugs per issues #34 and #35
