### 2018-04-04 acbd433
Fix insight addin bugs per issues #34 and #35 (#36)
