### 2018-12-20 9c4d0a8
Insight addin form validation (#37)
* Inline dw_set_config()

* Space/tab options in .rproj

* Add more user-friendly form validation and error reporting
