### 2018-12-21 47537b2
Improve docs and prepare 1.2.3 release
