+++ DESCRIPTION
@@ -2,7 +2,8 @@ Package: data.world
 Title: Functions and Add-Ins for Working with 'data.world' Data Sets and Projects
 Version: 1.2.3
 Authors@R: c(
-    person("Rafael", "Pereira", email = "rafael.pereira@data.world", role = c("aut", "cre")),
+    person("Mark", "DiMarco", email = "mark.dimarco@data.world", role = c("aut", "cre")),
+    person("Rafael", "Pereira", email = "rafael.pereira@data.world", role = c("aut")),
     person("Triet", "Le", email = "triet.le@data.world", role = c("aut")),
     person("Bryon", "Jacob", email = "bryon.jacob@data.world", role = c("aut")),
     person("Scott", "Came", email = "scott@cascadia-analytics.com", role = c("aut")))
