### 2019-02-25 #133 @sckott
According to our policies https://ropensci.github.io/dev_guide/policies.html#review-process it's been more than 1 year since this has been on hold (holding label) and will now be closed.
